+++ src/components/ViewProfile/index.js
@@ -2,48 +2,12 @@ import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 import {ScrollView, View} from 'react-native';
 
-import Text from '../Text';
 import ProfileCard from '../ProfileCard';
 import TabMenu from '../TabMenu';
+import {Info, NoInfo} from './Info';
 import {CHAT} from '../../constants/screens';
 
 import styles from './styles';
-
-const Info = ({shynee}) => <View style={styles.info}>
-  {shynee.name ? <View style={styles.infoItem}>
-    <Text style={styles.itemName}>NAME</Text>
-    <Text style={styles.itemValue}>{shynee.name}</Text>
-  </View> : null}
-  {shynee.dob ? <View style={styles.infoItem}>
-    <Text style={styles.itemName}>DATE OF BIRTH</Text>
-    <Text style={styles.itemValue}>{shynee.dob}</Text>
-  </View> : null}
-  {shynee.gender ? <View style={styles.infoItem}>
-    <Text style={styles.itemName}>SEX</Text>
-    <Text style={styles.itemValue}>{shynee.gender}</Text>
-  </View> : null}
-  {shynee.interests ? <View style={styles.infoItem}>
-    <Text style={styles.itemName}>INTERESTS</Text>
-    <View style={styles.interests}>{
-      ['ablab', 'lalab'].map((interest, index) => 
-        <View key={index} style={styles.interestContainer}>
-          <Text style={styles.interest}>{interest}</Text>
-        </View>
-      )}
-    </View>
-  </View> : null}
-  {shynee.personalInfo ? <View style={styles.infoItem}>
-    <Text style={styles.itemName}>ABOUT ME</Text>
-    <Text style={styles.itemValue}>Lorem ipsum dolor sit, amet consectetur adipisicing elit. Praesentium optio, fugiat eos rem suscipit magnam adipisci illum, quod a omnis possimus distinctio explicabo in earum minus sit atque reprehenderit similique!</Text>
-  </View> : null}
-</View>;
-
-const NoInfo = () => <View style={styles.noInfoContainer}>
-  <View style={styles.noInfo}>
-    <Text style={styles.noInfoPity}>Unfontunaley, no info about Florence Adams is public.</Text>
-    <Text style={styles.noInfoSuggestion}>You can still have a nice talk!</Text>
-  </View>
-</View>;
 
 class ViewProfile extends PureComponent {
   constructor(props) {
@@ -85,10 +49,6 @@ class ViewProfile extends PureComponent {
 ViewProfile.propTypes = {
   navigation: PropTypes.object,
   dispatch: PropTypes.func,
-  shynee: PropTypes.object,
-};
-
-Info.propTypes = {
   shynee: PropTypes.object,
 };
